[Feat/#1] AuthLayout 컴포넌트화
