+++ frontend/src/components/Auth/AuthLayout.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+interface AuthLayoutProps {
+  children: React.ReactNode;
+  title: string;
+}
+
+const AuthLayout = ({ children, title }: AuthLayoutProps) => {
+  return (
+    <div className="flex justify-center items-center min-h-screen bg-gray-100">
+      <div className="max-w-md w-full bg-white p-10 rounded-lg shadow-lg">
+        <h2 className="text-center text-2xl font-extrabold mb-6">{title}</h2>
+        {children}
+      </div>
+    </div>
+  );
+};
+
+export default AuthLayout;
